requirements.txt
smolagents
google-generativeai
